move letification into separate class
The letification algorithm is the most complex aspect of the conversion to
SMT2, and is relatively independent of the rest.  It may eventually have
uses elsewhere.
