+++ src/solvers/Makefile
@@ -181,6 +181,7 @@ SRC = $(BOOLEFORCE_SRC) \
       sat/dimacs_cnf.cpp \
       sat/pbs_dimacs_cnf.cpp \
       sat/resolution_proof.cpp \
+      smt2/letify.cpp \
       smt2/smt2_conv.cpp \
       smt2/smt2_dec.cpp \
       smt2/smt2_format.cpp \
