+++ src/solvers/smt2/smt2_conv.cpp
@@ -4713,85 +4713,6 @@ void smt2_convt::find_symbols_rec(
   }
 }
 
-exprt smt2_convt::letify(exprt &expr)
-{
-  seen_expressionst map;
-  std::vector<exprt> let_order;
-
-  collect_bindings(expr, map, let_order);
-
-  return letify_rec(expr, let_order, map, 0);
-}
-
-exprt smt2_convt::letify_rec(
-  exprt &expr,
-  std::vector<exprt> &let_order,
-  const seen_expressionst &map,
-  std::size_t i)
-{
-  if(i>=let_order.size())
-    return substitute_let(expr, map);
-
-  exprt current=let_order[i];
-  INVARIANT(
-    map.find(current) != map.end(), "expression should have been seen already");
-
-  if(map.find(current)->second.count < LET_COUNT)
-    return letify_rec(expr, let_order, map, i+1);
-
-  return let_exprt(
-    map.find(current)->second.let_symbol,
-    substitute_let(current, map),
-    letify_rec(expr, let_order, map, i + 1));
-}
-
-void smt2_convt::collect_bindings(
-  exprt &expr,
-  seen_expressionst &map,
-  std::vector<exprt> &let_order)
-{
-  seen_expressionst::iterator entry = map.find(expr);
-
-  if(entry != map.end())
-  {
-    let_count_idt &count_id = entry->second;
-    ++(count_id.count);
-    return;
-  }
-
-  // do not letify things with no children
-  if(expr.operands().empty())
-    return;
-
-  Forall_operands(it, expr)
-    collect_bindings(*it, map, let_order);
-
-  INVARIANT(
-    map.find(expr) == map.end(), "expression should not have been seen yet");
-
-  symbol_exprt let=
-    symbol_exprt("_let_"+std::to_string(++let_id_count), expr.type());
-
-  map.insert(std::make_pair(expr, let_count_idt(1, let)));
-
-  let_order.push_back(expr);
-}
-
-exprt smt2_convt::substitute_let(
-  exprt &expr,
-  const seen_expressionst &map)
-{
-  if(expr.operands().empty())
-    return expr;
-
-  let_visitort lv(map);
-
-  Forall_operands(it, expr)
-    it->visit(lv);
-
-  return expr;
-}
-
 std::size_t smt2_convt::get_number_of_solver_calls() const
 {
   return number_of_solver_calls;
